fix db, delete unique constraint

+++ main_bot/database/location_table.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Column, Integer, DECIMAL, ARRAY, String
+from sqlalchemy import Column, BigInteger, DECIMAL, ARRAY, String
 from .user_table import RelateUser
 
 
 class LocationUser(RelateUser):
     __tablename__ = 'user_loc'
-    id = Column(Integer(), nullable=False, autoincrement=True, primary_key=True)
+    id = Column(BigInteger(), nullable=False, autoincrement=True, primary_key=True)
     lon = Column(DECIMAL, nullable=False)
     lat = Column(DECIMAL, nullable=False)
     address = Column(String(), nullable=False)
