+++ main_bot/database/user_images_table.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, TIMESTAMP, String
+from sqlalchemy import Column, BigInteger, TIMESTAMP, String
 
 from .user_table import RelateUser
 
 
 class UserImages(RelateUser):
     __tablename__ = "user_images"
-    id = Column(Integer(), nullable=False, autoincrement=True, primary_key=True)
+    id = Column(BigInteger(), nullable=False, autoincrement=True, primary_key=True)
     url = Column(String(), nullable=False)
     create_at = Column(TIMESTAMP, nullable=False)
